test: protocol schema CI check; first batch of structs (#11855)
Adding "Protocol Schema" check to CI - you can see it in the list of
checks. This test will ensure that structs involved in protocol or saved
to DB doesn't accidentally change, see #11755. If that's not the case,
it will output list of impacted structs and suggest to change
`protocol_schema.toml` file.

Example of output in such case:
```
Loaded 123 structs
New struct: FlatStateValue with hash 83834662
New struct: ValueRef with hash 2322946441
New TOML file written to: ./target/protocol_schema.toml
Please review the changes and copy the file to protocol_schema.toml if they are correct.
```

Enabling it for first group of structs, mainly from core/primitives, to
check that it works. I will add remaining structs later.

The change is mainly renaming ProtocolStruct with ProtocolSchema as I
see it a better fit now, and add ProtocolSchema to the list of derives.
Note that if `protocol_schema` and `enable_const_type_id` are not
enabled, this derive will be no-op, so it won't impact regular code
generation.

cc @nagisa

---------

Co-authored-by: Simonas Kazlauskas <github@kazlauskas.me>