feat: Add temporary DB to replay-archival command to start eliminating unnecessary columns (#11876)
The goal of the `neard replay-archival` command is to demonstrate that
we can replay the chain from the genesis with minimal amount of data.

We previously run the command with read-only SplitDB. But to show that,
for example, we can derive data such as BlockInfo and EpochInfo on the
fly while replaying the blocks/chunks, we need to let components such as
EpochManager to generate these, write to a temporary DB (by calling
`add_validator_proposals`) and then read from there (instead of SplitDB
as before). As a step towards this, we add a new DB special to this
command: ReplayDB. It internally uses SplitDB for reading from archival
data, and a TestDB for writing temporary objects. ReplayDB is inspired
from MixedDB, but not exactly have the same functionality.

Then we explicitly enumerate the archival columns and make sure we only
read these columns from SplitDB and use the TestDB for other data
generate while replaying the blocks/chunks.

We will use this mechanism as we eliminate more columns from the
archival columns in future PRs.