test: Refactor/simplify view_requests_to_archival_node test and add more test cases (#11879)
The test `view_requests_to_archival_node` is used to exercise different
variations of the view-client requests on an archival node (so to view
historical information). This will be used to make sure as we do not
break view requests as we work with archival data (eg. eliminate some
columns from cold db).

In this PR, we simplify the calling of view methods; previously we used
a `view_client` library, in this PR we remove it and use a generic
`send` method instead. We also add `ViewClientTester` to simplify
function calls (use struct members instead of passing references along
the call stack).

Then we add more test cases, eg. execution outcome and state changes. We
will continue adding more test cases in future PRs.