feat: resharding tool (#11849)
The PR adds a new tool to perform resharding on demand. This's based on
a draft from @wacban with some adjustment to fix compile and runtime
errors.

The resharding is based on historical data, so this command is expected
to work only on archival nodes.

I'm still in the process of testing the procedure on a real node.

WIP:
- [x] The tool writes into a new DB. How to merge changes back in the
main database? Solution: add flag to write missing trie nodes directly
into the cold DB