feat: Introduce mechanism to load EpochConfigs for production chains from JSON files per version (#11896)
Issue: #11265

We introduce a mechanism for storing `EpochConfigs` for
mainnet/testnet/mocknet (forknet) in JSON files for various protocol
versions and retrieving them from a new library called
`EpochConfigStore`. The library parses JSON files containing the epoch
configs and loads them in a map from (chain-id x version) to
`EpochConfig`. The library is accessed from `AllEpochConfig` when
`use_production_config` is set to true AND outside of tests (for this we
explicitly set test overrides to default even if there is no explicit
override to distinguish cases where a test uses a mainnet/testnet chanin
id with a custom `EpochConfig`).

We change the existing `for_protocol_version` method of `AllEpochConfig`
to `generate_epoch_config` to indicate that it generates a config in
code. The new version of `for_protocol_version` internall calls the
config store or `generate_epoch_config` based on whether the config
store is initialized.

We will need to follow-up with more changes to this PR since there are
tests that use mainnet/testnet chain id with custom `EpochConfig` which
conflicts with the config store (which is supposed to use production
configs for these chains). Our eventual goal is to remove
`generate_epoch_config` and only have (1) explicitly documented
production configs in JSON files or (2) custom-built configs in test but
not using the hard-coded changes.