fix: Fix serde deserialization for FinalExecutionOutcomeViewEnum (#11905)
Discussion:
https://near.zulipchat.com/#narrow/stream/297873-node/topic/An.20issue.20with.20FinalExecutionOutcomeView

We had to switch the order of the variants because
`FinalExecutionOutcome` is a subset of
`FinalExecutionOutcomeWithReceipt` which means serde could never resolve
to the latter.

Borsh attributes were dropped as per discussion. While technically a
breaking change, it seems like this enum is only intended to be used
with JSON-RPC endpoint. Therefore, we think it's unlikely that someone
serialized it with borsh.