feat(mirror): add a --mirror-db-path flag (#11871)
The mirror run command opens a separate rocksdb database to store data
related to the traffic we're sending. Before this change we open it
inside of the target home directory's data dir. This is okay but it
makes it error prone to replace the target home directory. For example,
when generating traffic for statelessnet, we sometimes wanted to replace
the data dir disk with a recent snapshot. Moving the db path to a
different directory allows doing this without touching the mirror db.
Here we add this --mirror-db-path flag to point to this different path,
and if it's not present, we put it in the old place for backwards
compatibility