[TestLoop] Add a test for syncing, and necessary TestLoop enhancements. (#11892)
Several improvements:

* Allow a new TestLoop to be constructed with reused Store's and
TempDir.
* Allow arbitrary override of node config in the testloop builder (this
isn't used in this PR but will be later).
* Networking code no longer uses the very awkward route back global
static map. It instead uses a shared state. This also allows us to
support routing not just based on AccountId, but also PeerId (which is
needed for e.g. BlockHeadersRequest).
* Add ViewClientActor support to the test loop networking layer, which
is needed for answering queries e.g. BlockHeadersRequest.
* In the log visualizer UI, support log files that contain multiple
testloop runs. When visualizing such a file, the UI will first ask you
which run to visualize.

Add a test for the scenario of bootstrapping a new node using header,
state, and block sync (though the test doesn't dictate which sync method
must be used). This is both a useful test, and also a preparation to add
epoch sync later.