fix: Fix block/chunk production prometheus-metrics for kicked out validators (#11880)
Related
[thread](https://near.zulipchat.com/#narrow/stream/407237-core.2Fstateless-validation/topic/Non-validator.20node.20expected.20to.20produce.20chunks/near/456509732)

We update the prometheus metrics for block/chunk production of
validators by looping over the shards to record
expected/produced_by_shard metric here:
https://github.com/near/nearcore/blob/aaff26cfe0ababaa85ad259e701f07c7f797c79c/chain/client/src/info.rs#L506

But if a validator is kicked out in the previous epoch, we need to
remove the metrics (to keep only the metrics for the active validators
in the current epoch). In this case, the current code records an empty
shard id list for the kicked-out validator here:
https://github.com/near/nearcore/blob/aaff26cfe0ababaa85ad259e701f07c7f797c79c/chain/client/src/info.rs#L861

This prevents the loop to iterate over the shards and update the metric
per shard (while the metrics for aggregated expected/produced are
updated to 0).

To fix this, we explicitly create an enum called
`ValidatorProductionStatus` which indicates whether the validator is
active (`ValidatorProductionStatus::Validator`) or kicked out
(`ValidatorProductionStatus::Kickout`) in the current epoch. The old
code indicates this with the same struct `ValidatorProductionStats` with
0 values and empty vectors. Then for the kicked out validators, the code
loops over the shard ids in the current epoch to remove the metrics for
the validator.