fix: Revert changes to pipelined runtime implementation (#11951)
This reverts commit 781805d5075c7e8fffb1cbe3512df90c2d461a97

The commit above is causing issues in processing chunks both in testnet
and mainnet. The issue can be reproduced on an healthy node by applying
the commit and then running:

```
/home/ubuntu/neard view-state -t hot apply-range --start-index 125812312 --end-index 125812312 --shard-id 2 --storage trie-free sequential
```

Note: block 125812312 will sooner or later be GC'd so the repro steps
will need an archival node and cold storage.

Fix tested locally on a mainnet canary.