fix(epoch-manager): fix nondeterminism in EpochManager::compute_exempted_kickout() (#11949)
In this function, we compute and collect ratios of expected blocks and
chunks for all validators in the epoch, and we exempt the top validators
from being kicked based on the `validator_max_kickout_stake_perc` epoch
config field. We build this list from a hashmap, which has
nondeterministic iteration order, and then sort by this ratio. But if
there are many with the same ratio, then the ones we pick are not
deterministic because we don't also sort by account ID. Fix it by also
sorting by account ID