feat: Update neard replay-archive command to mark Incoming/Outgoing receipts non-archival (#11974)
We have been using `neard replay-archive` command to show that we can
replay the chain with the minimal set of columns.

In this PR, we iterate through the implementation of the command by
removing `IncomingReceipts` and `OutgoingReceipts` from the archival
columns, since we can replay the chain without depending on them. Note
that the code still derives and saves them in the store while replaying
the chain since we use them for the next blocks.