Remove BlockHeaderInfo  (#11971)
Today we first build `BlockHeaderInfo` from `BlockHeader` and then pass
`BlockHeaderInfo` to `EpochManager::add_validator_proposals`, which
internally builds `BuildInfo` and saves it.

I was first thinking that `BlockHeaderInfo` abstracts the information in
the header and uses as an API to the epoch manager. But the
`epoch_manager` crate already knows about the block headers and instead
we can directly use `BlockInfo` for the same purpose (which contains the
same information), so there is no need to build and intermediate
`BlockHeaderInfo`. Thus, we add a `from_header` method to `BlockInfo` to
directly build `BlockInfo` from the block header (same mappings from
header to the `BlockHeaderInfo` apply) and pass it to the epoch manager.

This is preparation for other work to be able to build and save a
`BlockInfo` directly from a block header (in replaying chain command).