fix: Fix Nayduck test one_val.py (#12018)
In this test, the network starts with 2 validators and one of them is
killed at the beginning and the network continues with one validator
only. Until the killed validator is kicked out, it will cause chunks
being missed (due to missing chunk production and validation) and this
will make the transactions fail. Thus, we change the test to wait until
the killed validator is not a current validator. Block production
continues meanwhile.