[network] Allow some network messages to be sent to self (#12022)
This PR builds the basis to enable the network layer to redirect network
messages to self.

Currently, any network message that is sent to self, i.e. sent to our
own account id, is dropped. This can lead to some complications in the
logic of dealing with sending chunk endorsement and partial witness
messages.

For example, we send the chunk endorsement to the next 5 block
producers, but since messages to self are dropped, we need to special
case in to check if we are the block producer and handle that case
separately.

For partial witness, we need to be careful about managing parts for
which we are the owner, i.e. bypass the
`PartialEncodedStateWitnessMessage` message and directly send
`PartialEncodedStateWitnessForwardMessage`.

In subsequent PRs, we would remove these special cases.

Changes
- Shift `receive_routed_message` from peer_actor to network_state
- Call `receive_routed_message` async in `send_message_to_account`
function if we are the recipient.
- Special case some routed message via `allow_sending_to_self` which we
allow to send to ourselves.