[stateless_validation] Allow sending chunk endorsement to self (#12023)
Follow up on https://github.com/near/nearcore/pull/12022

This PR removes the special case handling for sending chunk endorsement
to block producer where we are the block producer.

Most of the tests don't completely handle propagating the network
message so they had to be slightly modified.

Since we no longer have to pass the ChunkEndorsementTracker around, we
can get rid of the `Arc` around it in client and mutex around the LRU
cache in trackerV2