fix: Add setters for BlockHeader and move BlockHeaderV5 to nighly by  (#12043)
In https://github.com/near/nearcore/pull/12024, we introduced
`BlockHeaderV5` but could not move it to Nightly because some tests were
relying on using only a single header version. In other words, it was
not possible to have two different header versions in stable and nightly
builds.

To address that, we remove the `get_mut` function (which returns
`BlockHeaderV4` only) and add a bunch of setter functions, one per
field, which chooses the right header version to mutate.

Then we move the feature enabling the `BlockHeaderV5` to Nightly.