chore: Update replay-headers to report differences in rewards and kickouts (#12048)
We have been using the `neard replay-headers` command to simulate
validator rewards and kickouts computations for past epochs to check the
difference between original and replayed runs.

Submitting the new version of the command we have been using for
simulating the reward/kickout computations for chunk validators in the
context of #11900.

This version no longer assumes that the original and replayed values are
the same and prints the diffs in kickouts and rewards (stakes).