fix(config): don't initialize GCS state sync for localnet configs (#12055)
The config generated by `neard init` gives the same `.state_sync` field
for invocations without the `--chain-id` option set as invocations with
`--chain-id mainnet`. This default value indicates that the node should
try to download state parts from GCS. But this of course makes no sense
for localnet tests, where all state is local to the machine running the
test. Before this change, nodes in localnet tests that need to state
sync (because for example a chunk producer has been kicked and the
assignments across shards has changed) actually attempt to download
state parts from GCS, and of course fail to do so.

So fix it by only setting this default value when `--chain-id mainnet`
or `--chain-id testnet` is given, and not setting it otherwise. Then we
add an `extra_state_dumper` option to the pytest `start_cluster()` and
`init_cluster()` functions that will add an extra node to the test that
will dump state to a local directory, and set the state sync configs in
the other nodes to point to that directory